+++ Material/Material3D/Sand/DafaliasManzari.cpp
@@ -1,0 +1,336 @@
+/*******************************************************************************
+ * Copyright (C) 2017-2021 Theodore Chang
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ ******************************************************************************/
+
+// ReSharper disable IdentifierTypo
+// ReSharper disable StringLiteralTypo
+#include "DafaliasManzari.h"
+#include <Toolbox/tensorToolbox.h>
+
+const span DafaliasManzari::sb(1, 6);
+const span DafaliasManzari::sk(2, 7);
+const span DafaliasManzari::sl(8, 13);
+const span DafaliasManzari::sm(14, 19);
+const mat DafaliasManzari::unit_dev_tensor = tensor::unit_deviatoric_tensor4();
+
+DafaliasManzari::DafaliasManzari(const unsigned T, const double G0, const double NU, const double AC, const double LC, const double E0, const double XI, const double M, const double H0, const double CH, const double NB, const double A, const double ND, const double ZM, const double CZ, const double PC, const double GR, const double R)
+	: DataDafaliasManzari{fabs(G0), fabs(NU), fabs(AC), fabs(LC), fabs(E0), fabs(XI), fabs(M), fabs(H0), fabs(CH), fabs(NB), -fabs(A), fabs(ND), fabs(ZM), fabs(CZ), -fabs(PC), fabs(GR)}
+	, Material3D(T, R) { access::rw(tolerance) = 1E-12; }
+
+void DafaliasManzari::initialize(const shared_ptr<DomainBase>&) {
+	trial_stiffness = current_stiffness = initial_stiffness = tensor::isotropic_stiffness(gi * (2. + 2. * poissons_ratio), poissons_ratio);
+
+	initialize_history(12);
+}
+
+unique_ptr<Material> DafaliasManzari::get_copy() { return make_unique<DafaliasManzari>(*this); }
+
+double DafaliasManzari::get_parameter(const ParameterType P) const {
+	if(ParameterType::DENSITY == P) return density;
+	if(ParameterType::ELASTICMODULUS == P || ParameterType::YOUNGSMODULUS == P || ParameterType::E == P) return gi * (2. + 2. * poissons_ratio);
+	if(ParameterType::SHEARMODULUS == P || ParameterType::G == P) return gi;
+	if(ParameterType::POISSONSRATIO == P) return poissons_ratio;
+	return 0.;
+}
+
+int DafaliasManzari::update_trial_status(const vec& t_strain) {
+	incre_strain = (trial_strain = t_strain) - current_strain;
+
+	if(norm(incre_strain) <= tolerance) return SUANPAN_SUCCESS;
+
+	const auto current_p = tensor::mean(current_stress);
+	const auto current_s = tensor::dev(current_stress);
+	const auto incre_ev = tensor::trace(incre_strain);
+	const vec incre_ed = unit_dev_tensor * incre_strain;
+
+	// assume no plasticity
+	// compute trial stress
+
+	auto p = current_p + pr * gi * incre_ev;
+	vec s = current_s + 2. * gi * incre_ed;
+
+	const auto void_ratio = e0 + (1. + e0) * tensor::trace(trial_strain);
+	const auto v_term_a = pow(2.97 - void_ratio, 2.) / (1. + void_ratio);
+	const auto v_term_b = (void_ratio * (void_ratio + 2.) - 14.7609) * pow(1. + void_ratio, -2.) * (1. + e0);
+
+	double g, pgpe, pgpp;
+
+	vec residual(7), incre;
+	mat jacobian(7, 7, fill::eye);
+
+	auto counter = 0u;
+	auto ref_error = 1.;
+
+	while(true) {
+		if(max_iteration == ++counter) return SUANPAN_FAIL;
+
+		const auto sqrt_term = shear_modulus * sqrt(std::max(datum::eps, pc * p));
+
+		g = sqrt_term * v_term_a;
+
+		if(g > gi) {
+			pgpe = sqrt_term * v_term_b;
+			pgpp = .5 * g / p;
+		}
+		else {
+			g = gi;
+			pgpe = 0.;
+			pgpp = 0.;
+		}
+
+		residual(sa) = p - current_p - pr * g * incre_ev;
+		residual(sb) = s - current_s - 2. * g * incre_ed;
+
+		jacobian(sa, sa) = 1. - pr * incre_ev * pgpp;
+		jacobian(sb, sa) = -2. * pgpp * incre_ed;
+
+		if(!solve(incre, jacobian, residual)) return SUANPAN_FAIL;
+
+		auto error = norm(residual);
+		if(1 == counter) ref_error = std::max(1., error);
+		suanpan_debug("DafaliasManzari local elastic iteration error: %.5E.\n", error /= ref_error);
+		if(error <= tolerance) break;
+
+		p -= incre(sa);
+		s -= incre(sb);
+	}
+
+	// check if yield
+
+	const vec current_alpha(&current_history(0), 6, false, true);
+
+	vec eta = s + p * current_alpha;
+	auto norm_eta = tensor::stress::norm(eta);
+
+	if(norm_eta + m * p < 0.) {
+		trial_stress = s + p * tensor::unit_tensor2;
+
+		mat left(7, 6), right;
+
+		left.row(sa) = pr * (incre_ev * pgpe + g) * tensor::unit_tensor2.t();
+		left.rows(sb) = 2. * pgpe * incre_ed * tensor::unit_tensor2.t() + 2. * g * unit_dev_tensor;
+
+		if(!solve(right, jacobian, left)) return SUANPAN_FAIL;
+
+		trial_stiffness = right.rows(sb) + tensor::unit_tensor2 * right.row(sa);
+
+		return SUANPAN_SUCCESS;
+	}
+
+	// yield function violated
+
+	const vec current_z(&current_history(6), 6, false, true);
+
+	trial_history = current_history;
+	vec alpha(&trial_history(0), 6, false, true);
+	vec z(&trial_history(6), 6, false, true);
+
+	residual.set_size(20);
+	jacobian.set_size(20, 20);
+
+	counter = 0u;
+
+	vec n, zz, aabmn;
+	auto gamma = 0.;
+	double padpe, pabpe, d, pdpe, b0, pb0pe;
+
+	while(true) {
+		if(max_iteration == ++counter) return SUANPAN_FAIL;
+
+		auto sqrt_term = shear_modulus * sqrt(std::max(datum::eps, pc * p));
+
+		g = sqrt_term * v_term_a;
+
+		if(g > gi) {
+			pgpe = sqrt_term * v_term_b;
+			pgpp = .5 * g / p;
+		}
+		else {
+			g = gi;
+			pgpe = 0.;
+			pgpp = 0.;
+		}
+
+		const auto power_term = lc * pow(std::max(datum::eps, p / pc), xi);
+		const auto psi = void_ratio - e0 + power_term;
+		const auto ppsipp = xi * power_term / p;
+
+		const auto ad = ac * exp(nd * psi);
+		const auto ab = ac * exp(-nb * psi);
+		const auto adm = ad - m;
+		const auto abm = ab - m;
+
+		padpe = nd * ad;
+		pabpe = -nb * ab;
+		const auto padpp = padpe * ppsipp;
+		const auto pabpp = pabpe * ppsipp;
+
+		eta = s + p * alpha;
+		norm_eta = tensor::stress::norm(eta);
+
+		n = eta / norm_eta;
+		const vec unit_n = n % tensor::stress::norm_weight;
+		const vec unit_alpha = alpha % tensor::stress::norm_weight;
+		const auto alpha_n = dot(n, unit_alpha);
+		aabmn = alpha - adm * n;
+
+		const auto& etap = alpha_n;
+		const auto& etas = unit_n;
+		const vec etaa = p * etas;
+
+		const vec np = (alpha - alpha_n * n) / norm_eta;
+		const mat ns = (eye(6, 6) - n * unit_n.t()) / norm_eta;
+		const mat na = p * ns;
+
+		const vec unit_z = z % tensor::stress::norm_weight;
+		const auto zn = dot(n, unit_z);
+
+		d = a * (adm - alpha_n);
+
+		double pdpp;
+		rowvec pdps, pdpa, pdpz;
+		if(zn > 0.) {
+			const auto term_a = a * (adm - alpha_n);
+			const auto term_b = a * (1. + zn);
+
+			d *= 1. + zn;
+
+			pdpe = term_b * padpe * (1. + e0);
+			pdpp = term_a * dot(unit_z, np) + term_b * (padpp - dot(unit_alpha, np));
+			pdps = term_a * unit_z.t() * ns - term_b * unit_alpha.t() * ns;
+			pdpa = term_a * unit_z.t() * na - term_b * (unit_alpha.t() * na + unit_n.t());
+			pdpz = term_a * unit_n.t();
+		}
+		else {
+			pdpe = a * padpe * (1. + e0);
+			pdpp = a * (padpp - dot(unit_alpha, np));
+			pdps = -a * unit_alpha.t() * ns;
+			pdpa = -a * (unit_alpha.t() * na + unit_n.t());
+			pdpz.zeros(6);
+		}
+
+		sqrt_term = shear_modulus * h0 * sqrt(std::max(datum::eps, pc / p));
+		b0 = sqrt_term * (1. - ch * void_ratio);
+		pb0pe = -ch * sqrt_term * (1. + e0);
+		const auto pb0pp = -.5 * b0 / p;
+
+		residual(si) = norm_eta + m * p;
+		residual(sj) = p - current_p + pr * g * (gamma * d - incre_ev);
+		residual(sk) = s - current_s + 2. * g * (gamma * n - incre_ed);
+		residual(sl) = alpha - current_alpha + gamma * b0 * aabmn;
+
+		jacobian(si, si) = 0.;
+		jacobian(si, sj) = etap + m;
+		jacobian(si, sk) = etas.t();
+		jacobian(si, sl) = etaa.t();
+		jacobian(si, sm).zeros();
+
+		jacobian(sj, si) = d * pr * g;
+		jacobian(sj, sj) = 1. + pr * pgpp * (gamma * d - incre_ev) + pr * g * gamma * pdpp;
+		jacobian(sj, sk) = gamma * pr * g * pdps;
+		jacobian(sj, sl) = gamma * pr * g * pdpa;
+		jacobian(sj, sm) = gamma * pr * g * pdpz;
+
+		jacobian(sk, si) = 2. * g * n;
+		jacobian(sk, sj) = 2. * pgpp * (gamma * n - incre_ed) + 2. * g * gamma * np;
+		jacobian(sk, sk) = eye(6, 6) + 2. * g * gamma * ns;
+		jacobian(sk, sl) = 2. * g * gamma * na;
+		jacobian(sk, sm).zeros();
+
+		jacobian(sl, si) = b0 * aabmn;
+		jacobian(sl, sj) = gamma * pb0pp * aabmn - gamma * b0 * (pabpp * n + abm * np);
+		jacobian(sl, sk) = -gamma * b0 * abm * ns;
+		jacobian(sl, sl) = (1. + gamma * b0) * eye(6, 6) - gamma * b0 * abm * na;
+		jacobian(sl, sm).zeros();
+
+		if(d > 0.) {
+			residual(sm) = (1. + cz * gamma * d) * z - current_z + cz * gamma * zm * n;
+
+			zz = zm * n + z;
+
+			jacobian(sm, si) = cz * d * zz;
+			jacobian(sm, sj) = cz * gamma * pdpp * zz + cz * gamma * d * zm * np;
+			jacobian(sm, sk) = cz * gamma * zz * pdps + cz * gamma * d * zm * ns;
+			jacobian(sm, sl) = cz * gamma * zz * pdpa + cz * gamma * d * zm * na;
+			jacobian(sm, sm) = (1. + cz * gamma * d) * eye(6, 6) + cz * gamma * zz * pdpz;
+		}
+		else {
+			residual(sm) = z - current_z;
+
+			jacobian(sm, si).zeros();
+			jacobian(sm, sj).zeros();
+			jacobian(sm, sk).zeros();
+			jacobian(sm, sl).zeros();
+			jacobian(sm, sm) = eye(6, 6);
+		}
+
+		if(!solve(incre, jacobian, residual)) return SUANPAN_FAIL;
+
+		auto error = norm(residual);
+		if(1 == counter) ref_error = std::max(1., error);
+		suanpan_debug("DafaliasManzari local plastic iteration error: %.5E.\n", error /= ref_error);
+		if(error <= tolerance) break;
+
+		gamma -= incre(si);
+		p -= incre(sj);
+		s -= incre(sk);
+		alpha -= incre(sl);
+		z -= incre(sm);
+	}
+
+	trial_stress = s + p * tensor::unit_tensor2;
+
+	mat left(20, 6), right;
+
+	left.row(si).zeros();
+	left.row(sj) = pr * pgpe * (gamma * d - incre_ev) * tensor::stress::norm_weight.t() + pr * g * (gamma * pdpe - 1.) * tensor::stress::norm_weight.t();
+	left.rows(sk) = 2. * pgpe * (gamma * n - incre_ed) * tensor::stress::norm_weight.t() - 2. * g * unit_dev_tensor;
+	left.rows(sl) = gamma * aabmn * pb0pe * tensor::stress::norm_weight.t() - gamma * b0 * n * pabpe * (1. + e0) * tensor::stress::norm_weight.t();
+
+	d > 0. ? left.rows(sm) = cz * gamma * zz * pdpe * tensor::stress::norm_weight.t() : left.rows(sm).zeros();
+
+	if(!solve(right, jacobian, -left)) return SUANPAN_FAIL;
+
+	trial_stiffness = right.rows(sk) + tensor::unit_tensor2 * right.row(sj);
+
+	return SUANPAN_SUCCESS;
+}
+
+int DafaliasManzari::clear_status() {
+	current_strain.zeros();
+	current_stress.zeros();
+	current_history = initial_history;
+	current_stiffness = initial_stiffness;
+	return reset_status();
+}
+
+int DafaliasManzari::commit_status() {
+	current_strain = trial_strain;
+	current_stress = trial_stress;
+	current_history = trial_history;
+	current_stiffness = trial_stiffness;
+	return SUANPAN_SUCCESS;
+}
+
+int DafaliasManzari::reset_status() {
+	trial_strain = current_strain;
+	trial_stress = current_stress;
+	trial_history = current_history;
+	trial_stiffness = current_stiffness;
+	return SUANPAN_SUCCESS;
+}
+
+void DafaliasManzari::print() { suanpan_info("A Dafalias--Manzari sand model. doi: 10.1061/(ASCE)0733-9399(2004)130:6(622)\n"); }