+++ Material/MaterialParser.cpp
@@ -66,6 +66,7 @@ int create_new_material(const shared_ptr<DomainBase>& domain, istringstream& com
 	else if(is_equal(material_id, "ConcreteTsai")) new_concretetsai(new_material, command);
 	else if(is_equal(material_id, "CoulumbFriction")) new_coulumbfriction(new_material, command);
 	else if(is_equal(material_id, "Dhakal")) new_dhakal(new_material, command);
+	else if(is_equal(material_id, "DafaliasManzari")) new_dafaliasmanzari(new_material, command);
 	else if(is_equal(material_id, "Elastic1D")) new_elastic1d(new_material, command);
 	else if(is_equal(material_id, "Elastic2D")) new_elastic2d(new_material, command);
 	else if(is_equal(material_id, "Elastic3D")) new_isotropicelastic3d(new_material, command);
@@ -1566,6 +1567,21 @@ void new_expmises1d(unique_ptr<Material>& return_obj, istringstream& command) {
 	}
 
 	return_obj = make_unique<ExpMises1D>(tag, elastic_modulus, yield_stress, a, b, c, density);
+}
+
+void new_dafaliasmanzari(unique_ptr<Material>& return_obj, istringstream& command) {
+	unsigned tag;
+	if(!get_input(command, tag)) {
+		suanpan_error("new_dafaliasmanzari() requires a valid tag.\n");
+		return;
+	}
+
+	vec p(std::initializer_list<double>{125., .05, 1.25, .02, .9, .7, .01, 7., .9, 1.1, -.7, 3.5, 4., 6E2, -130., .2, 0.});
+
+	auto idx = 0;
+	while(!command.eof() && idx < 17) if(double para; get_input(command, para)) p(idx++) = para;
+
+	return_obj = make_unique<DafaliasManzari>(tag, p(0), p(1), p(2), p(3), p(4), p(5), p(6), p(7), p(8), p(9), p(10), p(11), p(12), p(13), p(14), p(15), p(16));
 }
 
 void new_flag01(unique_ptr<Material>& return_obj, istringstream& command) {